requirements.txt
certifi==2026.2.25
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.5
distlib==0.4.0
filelock==3.25.0
google-api-core==2.30.0
google-auth==2.48.0
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
identify==2.6.17
idna==3.11
iniconfig==2.3.0
nodeenv==1.10.0
packaging==26.0
platformdirs==4.9.2
pluggy==1.6.0
pre_commit==4.1.0
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pytest==8.3.4
python-discovery==1.1.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
ruff==0.9.2
urllib3==2.6.3
virtualenv==21.1.0
